Add generated .gitignore.
The .gitignore was generated with gitignore.io with the tags:
 * Linux
 * OSX
 * Windows
 * Java
 * Vim
The gitignore does not yet ignore the directory in which javadocs are contained.
This has proven to be useful as javadocs are a big clunky set of files that are
expensive to store with git. And by committing changes to javadocs graphs are
screwed up on GitHub.