Sort gitignore eclipse targets into new category..
Per request of @TheLocust3 some of the eclipse targets are sorted into another
category as they did not apply to eclipse in the .gitignore. The targets were
kept and most of them have remained unreviewed on whether or not we want them.
Changes can be expected to this file. Some of the targets that were moved to new
locations were:
 * .gradle things
 * swap files .swp (you can see some of these popping up when using vim)
 * Other things, that I don't know, they seem to be backup artifacts or
 something like .nib.