Add doc/ and docs/ target to .gitignore.
Just like the header of this commit, the only changes are that the targets /doc
and /docs are added to the gitignore. This is to prevent the javadocs we
generate to be even be considered for committing. We had many problems with
having javadocs being committed as they contributed to the graphs on github
being messed up and git blobs being huuuuuuuge! Anyway the docs can be generated
locally and there is at this time an online version of the javadocs.