Remove unneeded @Override annotation.
This removes the unneeded @Override annotation in MainWindowListener.java. @Override
should only be used when overriding the method of a superclass not for methods
implemented because of an interface. In this case, the method the annotation was
preceding was for an interface and therefore it is not needed and I removed it.