Add Socket constructor and complimentary createIO function.
This recent change was delayed due to the big merge and it got lost in the
sands of time (or something). This change was made because there was a need for
a way to use SocketWrapper with it's nice methods and enums and general
beautiful interface with a ServerSocket, which surprisingly is not an subclass
of a Socket but rather a whole new class derived from nothing. To add onto the
issues one in Java cannot cast an object into a subclass of that object for
some technical reasons.

And so I created two things:
 * A constructor
 * A createIO method

I noticed that the whole SocketWrapper was really just manipulating the input
and output streams and thus it wasn't interacting with itself, or it's
superclass beyond the constructors and createIO. So I made a constructor that
takes a Socket instance and then calls a special createIO private method, one
that itself takes a socket instance and creates input and output streams with
it.

Maybe my explanation was overkill.