Add JSON parsing and package and switch on command.
It's all theoretical at this point, wait for following commits for a working
version. At the time of development there was a bug in socketwraper and so this
shouldn't be used until that bug is resolved. The bug is that the first
character is ommitted from the message, this is because one character is loaded
when getMessage is first called from the input stream, this is to do checks on
whether or not the JSON is correct or not.

It should be noted that now the functionality of checking the JSON, the source
of this bug is now obsolete since org.json does that itself.

I added the json.org jar from maven central as that has some packages needed for
parsing JSON. I have yet to test the gradle builds yet and whether or not they
work since I have been working directly with the jar in my eclipse environment.