Fix bug of omitting first character.
There was this bug where the first character of the read message was being
omitted this is because getMessage() checks that the first character is an
open bracket if the message type is JSON. However, while it reads a character
from the stream, it doesn't add it to the returned message.

So in this commit I simply added it back into the returned string.